feat: drop support of Nextcloud 27
